wip: isolate queue dependency diagnostics
